# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  /** ok, or a description of why the store is unusable */
  database: string;
  /** Server local time the scheduler compares check-in times against */
  time: string;
  timezone: string;
  /** Which Twilio delivery path is configured, or why none is */
  outboundMessaging: string;
}

export interface Error {
  error: string;
}

export type GoalStatus = typeof GoalStatus[keyof typeof GoalStatus];


export const GoalStatus = {
  proposed: 'proposed',
  active: 'active',
  locked: 'locked',
  failed: 'failed',
  completed: 'completed',
} as const;

export interface Goal {
  id: number;
  name: string;
  successCondition: string;
  failureCondition: string;
  status: GoalStatus;
  createdAt: string;
  /** @nullable */
  lockedAt: string | null;
  /** @nullable */
  editUnlockAt: string | null;
  /** @nullable */
  checkInTime: string | null;
}

export interface GoalInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  successCondition: string;
  /** @minLength 1 */
  failureCondition: string;
  /** @nullable */
  checkInTime?: string | null;
}

export interface GoalUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  successCondition?: string;
  /** @minLength 1 */
  failureCondition?: string;
  /** @nullable */
  checkInTime?: string | null;
}

export type EventType = typeof EventType[keyof typeof EventType];


export const EventType = {
  check_in: 'check_in',
  failure: 'failure',
  consequence_issued: 'consequence_issued',
  consequence_done: 'consequence_done',
  note: 'note',
  escalation_sent: 'escalation_sent',
} as const;

export interface Event {
  id: number;
  /** @nullable */
  goalId: number | null;
  type: EventType;
  content: string;
  createdAt: string;
}

export type EventInputType = typeof EventInputType[keyof typeof EventInputType];


export const EventInputType = {
  check_in: 'check_in',
  failure: 'failure',
  consequence_issued: 'consequence_issued',
  consequence_done: 'consequence_done',
  note: 'note',
  escalation_sent: 'escalation_sent',
} as const;

export interface EventInput {
  /** @nullable */
  goalId?: number | null;
  type: EventInputType;
  /** @minLength 1 */
  content: string;
}

/**
 * One rung of the predefined consequence ladder, keyed to how long the slip lasted.
 */
export interface LadderTier {
  /**
     * Smallest slip duration, in minutes, that this rung applies to.
     * @minimum 1
     */
  slipMinutes: number;
  /** @minLength 1 */
  consequence: string;
}

export interface Ladder {
  tiers: LadderTier[];
}

export interface LadderInput {
  /** @minItems 1 */
  tiers: LadderTier[];
}

export type TestMessageInputChannel = typeof TestMessageInputChannel[keyof typeof TestMessageInputChannel];


export const TestMessageInputChannel = {
  sms: 'sms',
  call: 'call',
} as const;

export interface TestMessageInput {
  channel: TestMessageInputChannel;
  /** Override the destination. Defaults to ACCOUNTABILITY_USER_PHONE. */
  to?: string;
}

export interface TestMessageResult {
  delivered: boolean;
  /** Which delivery path was used */
  mode: string;
  detail: string;
}

export interface SlipInput {
  /**
     * How many minutes were lost.
     * @minimum 1
     * @maximum 1440
     */
  minutes: number;
  /** Optional factual description of what happened. */
  note?: string;
  /** @nullable */
  goalId?: number | null;
}

export interface SlipResult {
  minutes: number;
  failureEvent: Event;
  consequenceEvent: Event;
  tier: LadderTier | null;
}

export interface Contact {
  /** @nullable */
  id: number | null;
  name: string;
  phoneNumber: string;
  email: string;
  /** @nullable */
  consentConfirmedAt: string | null;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 7 */
  phoneNumber: string;
  email?: string;
}

export interface Dashboard {
  activeGoals: number;
  proposedGoals: number;
  overdueGoals: number;
  checkInsToday: number;
  recentEvents: Event[];
}

export type ChatTurnRole = typeof ChatTurnRole[keyof typeof ChatTurnRole];


export const ChatTurnRole = {
  user: 'user',
  partner: 'partner',
} as const;

export interface ChatTurn {
  role: ChatTurnRole;
  text: string;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
  /**
     * Recent turns of this conversation, oldest first, excluding the current message. Supplies conversational continuity only; stored goal and event state remains the source of truth for any fact.
     * @maxItems 20
     */
  history?: ChatTurn[];
}

export interface ChatResponse {
  message: string;
  providerConfigured: boolean;
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type ListEventsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};