# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: goals
    description: Accountability goals
  - name: events
    description: Factual accountability history
  - name: settings
    description: Accountability configuration
  - name: dashboard
    description: Overview data
  - name: chat
    description: Accountability partner chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get accountability overview
      responses:
        "200":
          description: Current overview and recent activity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List accountability goals
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Propose a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Proposed goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    get:
      operationId: getGoal
      tags: [goals]
      summary: Get a goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Edit a proposed or unlocked goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "409":
          description: Goal is locked or in its amendment cooldown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a proposed or unlocked goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "204":
          description: Deleted
        "409":
          description: Goal cannot be deleted yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /goals/{id}/lock:
    post:
      operationId: lockGoal
      tags: [goals]
      summary: Confirm and lock a proposed goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Active locked goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "409":
          description: Goal is not proposed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /goals/{id}/amend:
    post:
      operationId: requestGoalAmendment
      tags: [goals]
      summary: Start the 24-hour amendment cooldown
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Goal with amendment window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List factual accountability events
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Log an accountability event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Logged event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /slips:
    post:
      operationId: logSlip
      tags: [events]
      summary: Record a time slip and issue its predefined consequence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlipInput"
      responses:
        "201":
          description: Slip recorded with the applicable ladder rung
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlipResult"
  /ladder:
    get:
      operationId: getLadder
      tags: [settings]
      summary: Get the predefined consequence ladder
      responses:
        "200":
          description: Ladder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ladder"
    put:
      operationId: updateLadder
      tags: [settings]
      summary: Replace the predefined consequence ladder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LadderInput"
      responses:
        "200":
          description: Ladder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ladder"
  /selftest/message:
    post:
      operationId: sendTestMessage
      tags: [settings]
      summary: Send a test SMS or call to the configured user phone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestMessageInput"
      responses:
        "200":
          description: Delivery attempt result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestMessageResult"
        "503":
          description: Outbound messaging is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contact:
    get:
      operationId: getContact
      tags: [settings]
      summary: Get the accountability contact
      responses:
        "200":
          description: Contact or empty state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    put:
      operationId: updateContact
      tags: [settings]
      summary: Save the accountability contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contact/confirm:
    post:
      operationId: confirmContact
      tags: [settings]
      summary: Confirm the contact has consented
      responses:
        "200":
          description: Confirmed contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the accountability partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "503":
          description: LLM provider has not been configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    GoalId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      required:
        - status
        - database
        - time
        - timezone
        - outboundMessaging
      properties:
        status:
          type: string
        database:
          type: string
          description: "ok, or a description of why the store is unusable"
        time:
          type: string
          description: Server local time the scheduler compares check-in times against
        timezone:
          type: string
        outboundMessaging:
          type: string
          description: Which Twilio delivery path is configured, or why none is
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Goal:
      type: object
      required: [id, name, successCondition, failureCondition, status, createdAt, lockedAt, editUnlockAt, checkInTime]
      properties:
        id: { type: integer }
        name: { type: string }
        successCondition: { type: string }
        failureCondition: { type: string }
        status:
          type: string
          enum: [proposed, active, locked, failed, completed]
        createdAt: { type: string, format: date-time }
        lockedAt: { type: ["string", "null"], format: date-time }
        editUnlockAt: { type: ["string", "null"], format: date-time }
        checkInTime: { type: ["string", "null"] }
    GoalInput:
      type: object
      required: [name, successCondition, failureCondition]
      properties:
        name: { type: string, minLength: 1 }
        successCondition: { type: string, minLength: 1 }
        failureCondition: { type: string, minLength: 1 }
        checkInTime: { type: ["string", "null"] }
    GoalUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        successCondition: { type: string, minLength: 1 }
        failureCondition: { type: string, minLength: 1 }
        checkInTime: { type: ["string", "null"] }
    Event:
      type: object
      required: [id, goalId, type, content, createdAt]
      properties:
        id: { type: integer }
        goalId: { type: ["integer", "null"] }
        type:
          type: string
          enum: [check_in, failure, consequence_issued, consequence_done, note, escalation_sent]
        content: { type: string }
        createdAt: { type: string, format: date-time }
    EventInput:
      type: object
      required: [type, content]
      properties:
        goalId: { type: ["integer", "null"] }
        type:
          type: string
          enum: [check_in, failure, consequence_issued, consequence_done, note, escalation_sent]
        content: { type: string, minLength: 1 }
    LadderTier:
      type: object
      description: One rung of the predefined consequence ladder, keyed to how long the slip lasted.
      required: [slipMinutes, consequence]
      properties:
        slipMinutes:
          type: integer
          minimum: 1
          description: Smallest slip duration, in minutes, that this rung applies to.
        consequence:
          type: string
          minLength: 1
    Ladder:
      type: object
      required: [tiers]
      properties:
        tiers:
          type: array
          items: { $ref: "#/components/schemas/LadderTier" }
    LadderInput:
      type: object
      required: [tiers]
      properties:
        tiers:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/LadderTier" }
    TestMessageInput:
      type: object
      required: [channel]
      properties:
        channel:
          type: string
          enum: [sms, call]
        to:
          type: string
          description: Override the destination. Defaults to ACCOUNTABILITY_USER_PHONE.
    TestMessageResult:
      type: object
      required: [delivered, mode, detail]
      properties:
        delivered: { type: boolean }
        mode:
          type: string
          description: Which delivery path was used
        detail: { type: string }
    SlipInput:
      type: object
      required: [minutes]
      properties:
        minutes:
          type: integer
          minimum: 1
          maximum: 1440
          description: How many minutes were lost.
        note:
          type: string
          description: Optional factual description of what happened.
        goalId:
          type: ["integer", "null"]
    SlipResult:
      type: object
      required: [minutes, failureEvent, consequenceEvent, tier]
      properties:
        minutes: { type: integer }
        failureEvent: { $ref: "#/components/schemas/Event" }
        consequenceEvent: { $ref: "#/components/schemas/Event" }
        tier:
          oneOf:
            - $ref: "#/components/schemas/LadderTier"
            - type: "null"
    Contact:
      type: object
      required: [id, name, phoneNumber, email, consentConfirmedAt]
      properties:
        id: { type: ["integer", "null"] }
        name: { type: string }
        phoneNumber: { type: string }
        email: { type: string }
        consentConfirmedAt: { type: ["string", "null"], format: date-time }
    ContactInput:
      type: object
      required: [name, phoneNumber]
      properties:
        name: { type: string, minLength: 1 }
        phoneNumber: { type: string, minLength: 7 }
        email: { type: string }
    Dashboard:
      type: object
      required: [activeGoals, proposedGoals, overdueGoals, checkInsToday, recentEvents]
      properties:
        activeGoals: { type: integer }
        proposedGoals: { type: integer }
        overdueGoals: { type: integer }
        checkInsToday: { type: integer }
        recentEvents:
          type: array
          items: { $ref: "#/components/schemas/Event" }
    ChatTurn:
      type: object
      required: [role, text]
      properties:
        role:
          type: string
          enum: [user, partner]
        text: { type: string }
    ChatInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
        history:
          type: array
          maxItems: 20
          description: >-
            Recent turns of this conversation, oldest first, excluding the current
            message. Supplies conversational continuity only; stored goal and event
            state remains the source of truth for any fact.
          items: { $ref: "#/components/schemas/ChatTurn" }
    ChatResponse:
      type: object
      required: [message, providerConfigured]
      properties:
        message: { type: string }
        providerConfigured: { type: boolean }