# lib/api-zod/src/generated/types/chatTurn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatTurnRole } from './chatTurnRole';

export interface ChatTurn {
  role: ChatTurnRole;
  text: string;
}