# lib/api-zod/src/generated/types/ladderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LadderTier } from './ladderTier';

export interface LadderInput {
  /** @minItems 1 */
  tiers: LadderTier[];
}