# lib/api-zod/src/generated/types/ladderTier.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * One rung of the predefined consequence ladder, keyed to how long the slip lasted.
 */
export interface LadderTier {
  /**
     * Smallest slip duration, in minutes, that this rung applies to.
     * @minimum 1
     */
  slipMinutes: number;
  /** @minLength 1 */
  consequence: string;
}