# lib/api-zod/src/generated/types/slipInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SlipInput {
  /**
     * How many minutes were lost.
     * @minimum 1
     * @maximum 1440
     */
  minutes: number;
  /** Optional factual description of what happened. */
  note?: string;
  /** @nullable */
  goalId?: number | null;
}