# lib/api-zod/src/generated/types/slipResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from './event';
import type { LadderTier } from './ladderTier';

export interface SlipResult {
  minutes: number;
  failureEvent: Event;
  consequenceEvent: Event;
  tier: LadderTier | null;
}