# lib/api-zod/src/generated/types/testMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TestMessageInputChannel } from './testMessageInputChannel';

export interface TestMessageInput {
  channel: TestMessageInputChannel;
  /** Override the destination. Defaults to ACCOUNTABILITY_USER_PHONE. */
  to?: string;
}