# lib/api-zod/src/generated/types/testMessageResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TestMessageResult {
  delivered: boolean;
  /** Which delivery path was used */
  mode: string;
  detail: string;
}